refactor: Game abstract class 추가
