+++ src/main/java/chess/domain/state/Finished.java
@@ -1,16 +1,11 @@
 package chess.domain.state;
 
-import chess.domain.Result;
 import chess.domain.piece.Piece;
-import chess.domain.pieceinformations.TeamColor;
 import chess.domain.position.Position;
-import chess.domain.team.PieceSet;
-import chess.domain.team.Score;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.Map;
 
-public class Finished implements GameState {
+public class Finished extends Game {
     private final Map<Position, Piece> chessBoard;
 
     public Finished(Map<Position, Piece> chessBoard) {
@@ -20,28 +15,6 @@ public class Finished implements GameState {
     @Override
     public GameState move(Position source, Position target) {
         throw new UnsupportedOperationException("게임이 종료되었으므로 움직일 수 없습니다.");
-    }
-
-    @Override
-    public Result result(PieceSet black, PieceSet white) {
-        Map<TeamColor, Score> result = teamScores(black, white);
-
-        if (result.get(TeamColor.BLACK).compareTo(result.get(TeamColor.WHITE)) > 0) {
-            return new Result(result, TeamColor.BLACK);
-        }
-        if (result.get(TeamColor.BLACK).compareTo(result.get(TeamColor.WHITE)) < 0) {
-            return new Result(result, TeamColor.WHITE);
-        }
-
-        return new Result(result, TeamColor.NONE);
-    }
-
-
-    private Map<TeamColor, Score> teamScores(PieceSet black, PieceSet white) {
-        Map<TeamColor, Score> result = new HashMap<>();
-        result.put(TeamColor.BLACK, black.calculateScore());
-        result.put(TeamColor.WHITE, white.calculateScore());
-        return result;
     }
 
     @Override
