+++ src/main/java/chess/domain/state/Running.java
@@ -1,17 +1,13 @@
 package chess.domain.state;
 
-import chess.domain.Result;
 import chess.domain.piece.Blank;
 import chess.domain.piece.Piece;
 import chess.domain.pieceinformations.TeamColor;
 import chess.domain.position.Position;
-import chess.domain.team.PieceSet;
-import chess.domain.team.Score;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.Map;
 
-public class Running implements GameState {
+public class Running extends Game {
     private final Map<Position, Piece> chessBoard;
     private final TeamColor turn;
 
@@ -56,28 +52,6 @@ public class Running implements GameState {
         chessBoard.put(source, Blank.INSTANCE);
         chessBoard.put(target, startPiece);
         startPiece.changePosition(target);
-    }
-
-    @Override
-    public Result result(PieceSet black, PieceSet white) {
-        Map<TeamColor, Score> result = teamScores(black, white);
-
-        if (result.get(TeamColor.BLACK).compareTo(result.get(TeamColor.WHITE)) > 0) {
-            return new Result(result, TeamColor.BLACK);
-        }
-        if (result.get(TeamColor.BLACK).compareTo(result.get(TeamColor.WHITE)) < 0) {
-            return new Result(result, TeamColor.WHITE);
-        }
-
-        return new Result(result, TeamColor.NONE);
-    }
-
-
-    private Map<TeamColor, Score> teamScores(PieceSet black, PieceSet white) {
-        Map<TeamColor, Score> result = new HashMap<>();
-        result.put(TeamColor.BLACK, black.calculateScore());
-        result.put(TeamColor.WHITE, white.calculateScore());
-        return result;
     }
